+++ g.js
@@ -1,14 +1,22 @@
 'use strict';
 
-var Promise = require('bluebird')
+var Promise = require('native-or-bluebird')
+var caka = require('./index')
 
 
 module.exports = function(opts) {
-  var golem = Promise.promisify(require('./')(opts))
+  function cakaAsync(req, res) {
+    return new Promise(function(resolve, reject) {
+      caka(req, res, function next(err) {
+        if (err) reject(err)
+        else resolve()
+      })
+    })
+  }
 
   return function*(next) {
     try {
-      yield golem(this.request, this.response)
+      yield cakaAsync(this.request, this.response)
     }
     catch (e) {
       yield next